# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "AFN",
        "description": "AFN",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "tags": [
        {
            "name": "AFN",
            "description": "API AFN project"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
       
        "/api/v1/user/signup": {
          "post": {
            "tags": [
              "user router"
            ],
            "description": "user signup",
            "parameters": [
              {
                "name": "body",
                "in": "body",
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            ],
            "responses": {
              "201": {
                "description": "Signer Up successfully",
                "schema": {
                  "type":"array",
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
    
      
        "/api/v1/user/confirmation/{token}": {
          "post": {
            "tags": [
              "user router"
            ],
            "description": "user email confirmation",
            "parameters": [
              {
                "name": "token",
                "in": "path",
                "required": true,
                "description": "token from email account",
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "The account has been verified. Please log in.",
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        },
    
        "/api/v1/user/resend": {
          "post": {
            "tags": [
              "user router"
            ],
            "description": "resend confirmation email",
            "parameters": [
              {
                "name": "email",
                "in": "body",
                "required": true,
                "description": "email used in registration",
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "confirmation email resend succesfully!",
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        },
        "components": {
            "schemas": {
              "User": {
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "Warrior"
                  },
                 
                  "firstname": {
                    "type": "string",
                    "example": "user"
                  },
                  "lastname": {
                    "type": "string",
                    "example": "Gorba Koks"
                  },
                  "email": {
                    "type": "string",
                    "example": "strikers668@gmail.vom"
                  },
                  "password": {
                    "type": "string",
                    "example": "kaka"
                  }
                },
                "required": [
                  "username",
                  "firstname",
                  "lastname",
                  "email",
                  "password"
                ]
              }
            }
          },
            
    
    "definitions": {
        "User": {
            "properties": {
                "username": {
                    "type": "string",
                    "required": true
                },
                "firstname": {
                    "type": "string",
                    "required": true
                },
                "lastname": {
                    "type": "string",
                    "required": true
                },
                "email": {
                    "type": "string",
                    "required": true
                },
                "password": {
                    "type": "string",
                    "required": true
                },
                "isVerified":{
                    "type":"boolean"
                },
                "provider":{
                    "type":"string"
                }


            }
        },
        "Profile": {
            "properties": {
              
                "gender": {
                    "type": "string",
                    "enum":["female","male","prefer not to say"],
                    "required": true
                },
                "birthdate": {
                    "type": "date",
                    "required": true
                },
                "country": {
                    "type": "string",
                    "required": true
                },
                "nationality": {
                    "type": "string",
                    "required": true
                },
                "province": {
                    "type": "string",
                    "required": true
                },
                "district": {
                    "type": "string",
                    "required": true
                },
                "phone": {
                    "type": "string",
                    "required": true
                },
                "proffesion": {
                    "type": "string",
                    "required": true
                }
            }
        },

        "Business": {
            "properties": {
                "business_name": {
                    "type": "string",
                    "required": true
                },
                "logo": {
                    "type": "string",
                    "required": true
                },
                "country": {
                    "type": "string",
                    "required": true
                },
                "province": {
                    "type": "string",
                    "required": true
                },
                "district": {
                    "type": "string",
                    "required": true
                },
                "sector": {
                    "type": "string",
                    "required": true
                },
                "phone": {
                    "type": "string",
                    "required": true
                },
                "currency": {
                    "type": "string",
                    "default":"RWF",
                    "enum":["RWF","USD","EURO"]
                },
                "tin": {
                    "type": "string",
                    "required": true
                },
                "website": {
                    "type": "string",
                    "required": true
                },
                "workspace": {
                    "type": "string",
                    "required": true
                },
                "about_business": {
                    "type": "string",
                    "required": true
                }
            }
        },
        "Role": {
            "properties": {
                "name": {
                    "type": "string",
                    "enum":["employee","admin","disable"]
                       
                }
            }
        }
    }
}
}